1、两个整数相除想要得到小数必须强转为float

+++ BiuBike/app/src/main/java/com/biubike/service/RouteService.java
@@ -265,7 +265,7 @@ public class RouteService extends Service {
 
             if(totalDistance>1000) {
                 DecimalFormat df = new DecimalFormat("#.00");
-                showDistance = df.format(totalDistance/1000)+ "千米";
+                showDistance = df.format((float)totalDistance/1000)+ "千米";
             }
             else showDistance=totalDistance + "米";
             if(totalTime>60) {
